+++ src/Objects/Collection.php
@@ -8,21 +8,21 @@ use andrefelipe\Orchestrate\Query\KeyRangeBuilder;
  *
  * @link https://orchestrate.io/docs/apiref
  */
-class Collection extends AbstractList
+class Collection extends AbstractList implements CollectionInterface
 {
+    use Properties\CollectionTrait;
     use Properties\EventClassTrait;
     use Properties\ItemClassTrait;
     use Properties\AggregatesTrait;
 
     /**
-     * Constructs an item instance. A KeyValue or a custom class you set with
-     * setItemClass().
-     *
-     * @param string $key
-     * @param string $ref
-     *
-     * @return KeyValueInterface
+     * @param string $collection
      */
+    public function __construct($collection = null)
+    {
+        $this->setCollection($collection);
+    }
+
     public function item($key = null, $ref = null)
     {
         return $this->getItemClass()->newInstance()
@@ -32,17 +32,6 @@ class Collection extends AbstractList
             ->setHttpClient($this->getHttpClient());
     }
 
-    /**
-     * Constructs an event instance. An Event or a custom class you set with
-     * setEventClass().
-     *
-     * @param string $key
-     * @param string $type
-     * @param int $timestamp
-     * @param int $ordinal
-     *
-     * @return EventInterface
-     */
     public function event($key = null, $type = null, $timestamp = null, $ordinal = null)
     {
         return $this->getEventClass()->newInstance()
@@ -54,10 +43,6 @@ class Collection extends AbstractList
             ->setHttpClient($this->getHttpClient());
     }
 
-    /**
-     *
-     * @return Events
-     */
     public function events($key = null, $type = null)
     {
         return (new Events())
@@ -67,14 +52,6 @@ class Collection extends AbstractList
             ->setHttpClient($this->getHttpClient());
     }
 
-    /**
-     * Gets total item count of the Collection.
-     *
-     * May return zero if request was unsuccesful, in which case you can check
-     * the response with "getResponse" or the aliases "getStatusCode/getStatus".
-     *
-     * @return int|null Item count, or null on request failure.
-     */
     public function getTotalItems()
     {
         // makes a straight Search query for no results
@@ -94,16 +71,6 @@ class Collection extends AbstractList
         return null;
     }
 
-    /**
-     * Gets total event count of the Collection.
-     *
-     * May return zero if request was unsuccesful, in which case you can check
-     * the response with "getResponse" or the aliases "getStatusCode/getStatus".
-     *
-     * @param string $type Optionally restrict to a specific event type.
-     *
-     * @return int|null Event count, or null on request failure.
-     */
     public function getTotalEvents($type = null)
     {
         // makes a straight Search query for no results
@@ -156,7 +123,7 @@ class Collection extends AbstractList
     public function toArray()
     {
         $data = parent::toArray();
-        $data['kind'] = 'collection';
+        $data['kind'] = static::KIND;
 
         if ($this->getItemClass()->name !== self::$defaultItemClassName) {
             $data['itemClass'] = $this->getItemClass()->name;
@@ -171,19 +138,6 @@ class Collection extends AbstractList
         return $data;
     }
 
-    /**
-     * Gets a lexicographically ordered list of items contained in a collection,
-     * specified by the limit and key range parameters.
-     *
-     * If there are more results available, the pagination URL can be checked
-     * with getNextUrl/getPrevUrl, and queried with nextPage/prevPage methods.
-     *
-     * @param int $limit The limit of items to return. Defaults to 10 and max to 100.
-     * @param KeyRangeBuilder $range
-     *
-     * @return boolean Success of operation.
-     * @link https://orchestrate.io/docs/apiref#keyvalue-list
-     */
     public function get($limit = 10, KeyRangeBuilder $range = null)
     {
         // define request options
@@ -199,21 +153,11 @@ class Collection extends AbstractList
         return $this->isSuccess();
     }
 
-    /**
-     * Deletes a collection. Warning this will permanently erase all data within
-     * this collection and cannot be reversed!
-     *
-     * To prevent accidental deletions, provide the current collection name as
-     * the parameter. The collection will only be deleted if both names match.
-     *
-     * @return boolean Success of operation.
-     * @link https://orchestrate.io/docs/apiref#collections-delete
-     */
     public function delete($collectionName)
     {
         if ($collectionName === $this->getCollection(true)) {
 
-            $this->request('DELETE', $this->getCollection(), ['query' => ['force' => 'true']]);
+            $this->request('DELETE', $collectionName, ['query' => ['force' => 'true']]);
 
             if ($this->isSuccess()) {
                 $this->setResponseValues();
@@ -224,16 +168,6 @@ class Collection extends AbstractList
         return false;
     }
 
-    /**
-     * @param string $query
-     * @param string|array $sort
-     * @param string|array $aggregate
-     * @param int $limit
-     * @param int $offset
-     *
-     * @return boolean Success of operation.
-     * @link https://orchestrate.io/docs/apiref#search-collection
-     */
     public function search($query, $sort = null, $aggregate = null, $limit = 10, $offset = 0)
     {
         // define request options